requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-ckeditor==6.7.1
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
grep==0.3.2
gunicorn==23.0.0
idna==3.7
packaging==24.2
phonenumbers==8.13.52
pillow==11.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.9.0
